backend/requirements.txt
colorama==0.3.9
Django==3.1.7
filetype==1.0.9
imgviz==1.4.1
lxml==4.2.5
numpy==1.20.1
opencv_python==4.5.4.60
Pillow==8.4.0
pycocotools==2.0.3
PyMySQL==1.0.2
requests==2.24.0
termcolor==1.1.0
